LayoutTests/imported/w3c/web-platform-tests/2dcontext/drawing-rectangles-to-the-canvas/2d.strokeRect.transform-expected.txt

2d.strokeRect.transform
fillRect is affected by transforms
Actual output:
Expected output:


PASS fillRect is affected by transforms 
